Printing file import tree is now handled by reporter
